Reorganise kotlin files
