+++ JTime-android/src/main/java/com/ismail_s/jtime/android/Constants.kt
@@ -1,4 +1,4 @@
-package com.ismail_s.jtime.android.activity
+package com.ismail_s.jtime.android
 
 internal object Constants {
     @JvmField val MASJID_ID = "com.ismail_s.jtime.android.MASJID_ID"
