+++ JTime-android/src/main/java/com/ismail_s/jtime/android/MasjidRecyclerViewAdapter.kt
@@ -1,18 +1,17 @@
-package com.ismail_s.jtime.android.activity
+package com.ismail_s.jtime.android
 
 import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import com.ismail_s.jtime.android.MasjidPojo
-import com.ismail_s.jtime.android.R
+import com.ismail_s.jtime.android.pojo.MasjidPojo
 import org.jetbrains.anko.find
 
 /**
  * [RecyclerView.Adapter] that can display a [MasjidPojo].
  */
-class MyItemRecyclerViewAdapter(private val mValues: List<MasjidPojo>, private val mainActivity: MainActivity) : RecyclerView.Adapter<MyItemRecyclerViewAdapter.ViewHolder>() {
+class MasjidRecyclerViewAdapter(private val mValues: List<MasjidPojo>, private val mainActivity: MainActivity) : RecyclerView.Adapter<MasjidRecyclerViewAdapter.ViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
         val view = LayoutInflater.from(parent.context).inflate(R.layout.fragment_item, parent, false)
