+++ JTime-android/src/main/java/com/ismail_s/jtime/android/fragment/NearbyTimesFragment.kt
@@ -1,4 +1,4 @@
-package com.ismail_s.jtime.android.activity
+package com.ismail_s.jtime.android.fragment
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -9,7 +9,7 @@ import android.widget.TextView
 import com.ismail_s.jtime.android.CalendarFormatter.formatCalendarAsTime
 import com.ismail_s.jtime.android.R
 import com.ismail_s.jtime.android.RestClient
-import com.ismail_s.jtime.android.SalaahType
+import com.ismail_s.jtime.android.pojo.SalaahType
 import nl.komponents.kovenant.ui.failUi
 import nl.komponents.kovenant.ui.successUi
 import org.jetbrains.anko.*
